requirements.txt
tweepy
discord.py
python-telegram-bot
google-api-python-client
google-auth
google-auth-oauthlib
schedule
python-dotenv
requests